Move e2e helpers and update import statements
