+++ test/e2e/mock-helpers.js
