+++ test/edu-benefits/1990/01-all-fields.e2e.spec.js
@@ -1,6 +1,6 @@
-const E2eHelpers = require('../../util/e2e-helpers');
-const Timeouts = require('../../util/timeouts.js');
-const EduHelpers = require('../../util/edu-1990-helpers');
+const E2eHelpers = require('../../e2e/e2e-helpers');
+const Timeouts = require('../../e2e/timeouts.js');
+const EduHelpers = require('../../e2e/edu-1990-helpers');
 
 module.exports = E2eHelpers.createE2eTest(
   (client) => {
