+++ test/hca/03-submit-text.e2e.spec.js
@@ -1,6 +1,6 @@
-const E2eHelpers = require('../util/e2e-helpers');
-const Timeouts = require('../util/timeouts.js');
-const HcaHelpers = require('../util/hca-helpers.js');
+const E2eHelpers = require('../e2e/e2e-helpers');
+const Timeouts = require('../e2e/timeouts.js');
+const HcaHelpers = require('../e2e/hca-helpers.js');
 
 module.exports = E2eHelpers.createE2eTest(
   (client) => {
