+++ test/messaging/00-required.e2e.spec.js
@@ -1,7 +1,7 @@
-const E2eHelpers = require('../util/e2e-helpers');
-const Timeouts = require('../util/timeouts.js');
-const MessagingHelpers = require('../util/messaging-helpers');
-const LoginHelpers = require('../util/login-helpers');
+const E2eHelpers = require('../e2e/e2e-helpers');
+const Timeouts = require('../e2e/timeouts.js');
+const MessagingHelpers = require('../e2e/messaging-helpers');
+const LoginHelpers = require('../e2e/login-helpers');
 const selectDropdown = E2eHelpers.selectDropdown;
 
 module.exports = E2eHelpers.createE2eTest(
