+++ test/sanity-check/smoke-test.e2e.spec.js
@@ -1,8 +1,8 @@
 // TODO(awong): Convert to ES6
 
 // TODO(awong): Move all this into ES6 syntax.
-const E2eHelpers = require('../util/e2e-helpers');
-const Timeouts = require('../util/timeouts.js');
+const E2eHelpers = require('../e2e/e2e-helpers');
+const Timeouts = require('../e2e/timeouts.js');
 
 module.exports = E2eHelpers.createE2eTest(
   (client) => {
